Extract the BitMask support to a separate file
